Move: point 도메인 분리를 위해 도메인 계층 이동
