+++ src/main/kotlin/io/ticketaka/api/common/infrastructure/event/EventDispatcher.kt
@@ -1,8 +1,8 @@
 package io.ticketaka.api.common.infrastructure.event
 
 import io.ticketaka.api.common.domain.DomainEvent
-import io.ticketaka.api.reservation.domain.point.PointChargeEvent
-import io.ticketaka.api.reservation.domain.point.PointRechargeEvent
+import io.ticketaka.api.point.domain.PointChargeEvent
+import io.ticketaka.api.point.domain.PointRechargeEvent
 import io.ticketaka.api.reservation.infrastructure.event.PointChargeEventConsumer
 import io.ticketaka.api.reservation.infrastructure.event.PointRechargeEventConsumer
 import org.springframework.stereotype.Component
