+++ src/main/kotlin/io/ticketaka/api/point/application/IdempotentService.kt
@@ -1,9 +1,9 @@
 package io.ticketaka.api.point.application
 
 import io.ticketaka.api.common.infrastructure.IdempotentKeyGenerator
-import io.ticketaka.api.reservation.domain.point.Idempotent
-import io.ticketaka.api.reservation.domain.point.IdempotentRepository
-import io.ticketaka.api.reservation.domain.point.PointHistory
+import io.ticketaka.api.point.domain.Idempotent
+import io.ticketaka.api.point.domain.IdempotentRepository
+import io.ticketaka.api.point.domain.PointHistory
 import org.springframework.stereotype.Service
 import java.math.BigDecimal
