+++ src/main/kotlin/io/ticketaka/api/point/application/PaymentService.kt
@@ -1,8 +1,8 @@
 package io.ticketaka.api.point.application
 
 import io.ticketaka.api.point.application.dto.PaymentCommand
-import io.ticketaka.api.reservation.domain.payment.Payment
-import io.ticketaka.api.reservation.domain.payment.PaymentRepository
+import io.ticketaka.api.point.domain.payment.Payment
+import io.ticketaka.api.point.domain.payment.PaymentRepository
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Service
