+++ src/main/kotlin/io/ticketaka/api/point/application/PointBalanceService.kt
@@ -2,7 +2,7 @@ package io.ticketaka.api.point.application
 
 import io.ticketaka.api.point.application.dto.BalanceQueryModel
 import io.ticketaka.api.point.application.dto.RechargeCommand
-import io.ticketaka.api.reservation.domain.point.PointRechargeEvent
+import io.ticketaka.api.point.domain.PointRechargeEvent
 import io.ticketaka.api.user.application.TokenUserQueryService
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.retry.annotation.Backoff
