+++ src/main/kotlin/io/ticketaka/api/point/application/PointService.kt
@@ -1,7 +1,7 @@
 package io.ticketaka.api.point.application
 
-import io.ticketaka.api.reservation.domain.point.Point
-import io.ticketaka.api.reservation.domain.point.PointRepository
+import io.ticketaka.api.point.domain.Point
+import io.ticketaka.api.point.domain.PointRepository
 import io.ticketaka.api.user.domain.User
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
