+++ src/main/kotlin/io/ticketaka/api/point/domain/IdempotentRepository.kt
@@ -1,4 +1,4 @@
-package io.ticketaka.api.reservation.domain.point
+package io.ticketaka.api.point.domain
 
 interface IdempotentRepository {
     fun save(idempotent: Idempotent): Idempotent
