+++ src/main/kotlin/io/ticketaka/api/point/domain/Point.kt
@@ -1,4 +1,4 @@
-package io.ticketaka.api.reservation.domain.point
+package io.ticketaka.api.point.domain
 
 import io.ticketaka.api.common.domain.AbstractAggregateRoot
 import io.ticketaka.api.common.exception.BadClientRequestException
