+++ src/main/kotlin/io/ticketaka/api/point/domain/PointChargeEvent.kt
@@ -1,18 +1,24 @@
-package io.ticketaka.api.reservation.domain.point
+package io.ticketaka.api.point.domain
 
 import io.ticketaka.api.common.domain.DomainEvent
 import io.ticketaka.api.user.domain.User
 import java.math.BigDecimal
 import java.time.LocalDateTime
 
-data class PointRechargeEvent(
-    val userTsid: String,
+data class PointChargeEvent(
     val userId: Long,
     val pointId: Long,
+    val balance: BigDecimal,
     val amount: BigDecimal,
     val occurredOn: LocalDateTime,
 ) : DomainEvent {
-    constructor(user: User, point: Point, amount: BigDecimal) : this(user.tsid, user.getId(), point.getId(), amount, LocalDateTime.now())
+    constructor(user: User, point: Point, balance: BigDecimal, amount: BigDecimal) : this(
+        user.getId(),
+        point.getId(),
+        balance,
+        amount,
+        LocalDateTime.now(),
+    )
 
     override fun occurredOn(): LocalDateTime {
         return this.occurredOn
