+++ src/main/kotlin/io/ticketaka/api/point/domain/PointHistory.kt
@@ -1,8 +1,9 @@
-package io.ticketaka.api.reservation.domain.payment
+package io.ticketaka.api.point.domain
 
-import io.ticketaka.api.common.domain.AbstractAggregateRoot
 import io.ticketaka.api.common.infrastructure.tsid.TsIdKeyGenerator
 import jakarta.persistence.Entity
+import jakarta.persistence.EnumType
+import jakarta.persistence.Enumerated
 import jakarta.persistence.GeneratedValue
 import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
@@ -11,34 +12,38 @@ import java.math.BigDecimal
 import java.time.LocalDateTime
 
 @Entity
-@Table(name = "payments")
-class Payment(
+@Table(name = "point_histories")
+class PointHistory(
     val tsid: String,
-    val amount: BigDecimal,
-    val paymentTime: LocalDateTime,
+    @Enumerated(EnumType.STRING)
+    val transactionType: TransactionType,
     val userId: Long,
     val pointId: Long,
-) : AbstractAggregateRoot() {
+    val amount: BigDecimal,
+    val createTime: LocalDateTime = LocalDateTime.now(),
+) {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long? = null
 
-    init {
-        registerEvent(PaymentApprovalEvent(this, userId, pointId, amount))
+    enum class TransactionType {
+        RECHARGE,
+        CHARGE,
     }
 
     companion object {
         fun newInstance(
-            amount: BigDecimal,
             userId: Long,
             pointId: Long,
-        ): Payment {
-            return Payment(
-                tsid = TsIdKeyGenerator.next("pm"),
-                amount = amount,
+            amount: BigDecimal,
+            transactionType: TransactionType,
+        ): PointHistory {
+            return PointHistory(
+                tsid = TsIdKeyGenerator.next("ph"),
                 userId = userId,
-                paymentTime = LocalDateTime.now(),
                 pointId = pointId,
+                amount = amount,
+                transactionType = transactionType,
             )
         }
     }
