+++ src/main/kotlin/io/ticketaka/api/point/domain/PointRepository.kt
@@ -1,4 +1,4 @@
-package io.ticketaka.api.reservation.domain.point
+package io.ticketaka.api.point.domain
 
 import java.math.BigDecimal
