+++ src/main/kotlin/io/ticketaka/api/point/domain/payment/PaymentRepository.kt
@@ -1,4 +1,4 @@
-package io.ticketaka.api.reservation.domain.payment
+package io.ticketaka.api.point.domain.payment
 
 interface PaymentRepository {
     fun save(payment: Payment): Payment
