+++ src/main/kotlin/io/ticketaka/api/reservation/infrastructure/event/PointEventHandler.kt
@@ -1,8 +1,8 @@
 package io.ticketaka.api.reservation.infrastructure.event
 
 import io.ticketaka.api.common.domain.EventBroker
-import io.ticketaka.api.reservation.domain.point.PointChargeEvent
-import io.ticketaka.api.reservation.domain.point.PointRechargeEvent
+import io.ticketaka.api.point.domain.PointChargeEvent
+import io.ticketaka.api.point.domain.PointRechargeEvent
 import org.springframework.context.event.EventListener
 import org.springframework.stereotype.Component
