+++ src/main/kotlin/io/ticketaka/api/reservation/infrastructure/jpa/JpaIdempotentRepository.kt
@@ -1,6 +1,6 @@
 package io.ticketaka.api.reservation.infrastructure.jpa
 
-import io.ticketaka.api.reservation.domain.point.Idempotent
+import io.ticketaka.api.point.domain.Idempotent
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface JpaIdempotentRepository : JpaRepository<Idempotent, String> {
