+++ src/main/kotlin/io/ticketaka/api/reservation/infrastructure/jpa/JpaPaymentRepository.kt
@@ -1,6 +1,6 @@
 package io.ticketaka.api.reservation.infrastructure.jpa
 
-import io.ticketaka.api.reservation.domain.payment.Payment
+import io.ticketaka.api.point.domain.payment.Payment
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface JpaPaymentRepository : JpaRepository<Payment, Long>
