+++ src/main/kotlin/io/ticketaka/api/reservation/infrastructure/jpa/JpaPointHistoryRepository.kt
@@ -1,6 +1,6 @@
 package io.ticketaka.api.reservation.infrastructure.jpa
 
-import io.ticketaka.api.reservation.domain.point.PointHistory
+import io.ticketaka.api.point.domain.PointHistory
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface JpaPointHistoryRepository : JpaRepository<PointHistory, Long>
