+++ src/main/kotlin/io/ticketaka/api/reservation/infrastructure/persistence/PointRepositoryComposition.kt
@@ -1,7 +1,7 @@
 package io.ticketaka.api.reservation.infrastructure.persistence
 
-import io.ticketaka.api.reservation.domain.point.Point
-import io.ticketaka.api.reservation.domain.point.PointRepository
+import io.ticketaka.api.point.domain.Point
+import io.ticketaka.api.point.domain.PointRepository
 import io.ticketaka.api.reservation.infrastructure.jpa.JpaPointRepository
 import org.springframework.stereotype.Repository
 import java.math.BigDecimal
