+++ src/test/kotlin/io/ticketaka/api/common/domain/queue/TokenWaitingQueueTest.kt
@@ -1,6 +1,6 @@
 package io.ticketaka.api.common.domain.queue
 
-import io.ticketaka.api.reservation.domain.point.Point
+import io.ticketaka.api.point.domain.Point
 import io.ticketaka.api.user.domain.Token
 import io.ticketaka.api.user.domain.User
 import org.junit.jupiter.api.Assertions.assertTrue
