+++ src/test/kotlin/io/ticketaka/api/point/application/PointBalanceServiceTest.kt
@@ -2,7 +2,7 @@ package io.ticketaka.api.point.application
 
 import io.ticketaka.api.common.exception.BadClientRequestException
 import io.ticketaka.api.point.application.dto.RechargeCommand
-import io.ticketaka.api.reservation.domain.point.Point
+import io.ticketaka.api.point.domain.Point
 import io.ticketaka.api.user.application.TokenUserQueryService
 import io.ticketaka.api.user.domain.User
 import org.junit.jupiter.api.Assertions.assertEquals
